Factor out heavy dependencies for jira-issue-fetcher
