+++ ymir/jira_issue_fetcher/__init__.py
@@ -1,5 +1,1 @@
 """Jira issue fetcher for Ymir AI workflows."""
-
-from .jira_issue_fetcher import JiraIssueFetcher
-
-__all__ = ["JiraIssueFetcher"]
