+++ ymir/tools/constants.py
@@ -1,0 +1,5 @@
+import aiohttp
+
+AIOHTTP_TIMEOUT = aiohttp.ClientTimeout(total=30)
+
+BREWHUB_URL = "https://brewhub.engineering.redhat.com/brewhub"
