+++ ymir/tools/privileged/distgit.py
@@ -11,8 +11,8 @@ from beeai_framework.emitter import Emitter
 from beeai_framework.tools import StringToolOutput, Tool, ToolError, ToolRunOptions
 from pydantic import BaseModel, Field
 
-from ymir.common.constants import BREWHUB_URL
-from ymir.common.utils import KerberosError, init_kerberos_ticket
+from ymir.common.base_utils import KerberosError, init_kerberos_ticket
+from ymir.tools.constants import BREWHUB_URL
 
 SYNC_TIMEOUT = 1 * 60 * 60  # seconds
