+++ ymir/tools/unprivileged/commands.py
@@ -6,7 +6,7 @@ from beeai_framework.emitter import Emitter
 from beeai_framework.tools import JSONToolOutput, Tool, ToolError, ToolRunOptions
 from pydantic import BaseModel, Field
 
-from ymir.common.utils import run_subprocess
+from ymir.common.base_utils import run_subprocess
 
 TIMEOUT = 10 * 60  # seconds
 ELLIPSIZED_LINES = 200
