+++ ymir/tools/unprivileged/upstream_search.py
@@ -9,7 +9,7 @@ from beeai_framework.emitter import Emitter
 from beeai_framework.tools import JSONToolOutput, Tool, ToolError, ToolRunOptions
 from pydantic import BaseModel, Field
 
-from ymir.common.constants import AIOHTTP_TIMEOUT
+from ymir.tools.constants import AIOHTTP_TIMEOUT
 
 logger = logging.getLogger(__name__)
