+++ ymir/tools/unprivileged/upstream_tools.py
@@ -15,9 +15,9 @@ from beeai_framework.tools import (
 )
 from pydantic import BaseModel, Field
 
-from ymir.common.constants import AIOHTTP_TIMEOUT
-from ymir.common.utils import run_subprocess
+from ymir.common.base_utils import run_subprocess
 from ymir.common.validators import AbsolutePath
+from ymir.tools.constants import AIOHTTP_TIMEOUT
 
 
 class ExtractUpstreamRepositoryInput(BaseModel):
